Module_Azure_Function/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-storage-blob
numpy
ipywidgets
matplotlib
scikit-image
pydicom
plotly
chart_studio
